ctn2md/requirements.txt
acres==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
colorlog==6.9.0
configobj==5.0.9
configparser==7.1.0
crcmod==1.7
cryptography==44.0.0
dashscope==1.20.14
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
etelemetry==0.3.1
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
isodate==0.6.1
jieba==0.42.1
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
json_repair==0.30.3
llama-index-core==0.12.5
llama-parse==0.5.17
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.23.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
nltk==3.9.1
numpy==2.2.0
openai==1.57.2
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
propcache==0.2.1
prov==2.0.1
puremagic==1.28
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.3
pyflakes==3.2.0
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scipy==1.14.1
shapely==2.0.6
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
traits==6.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
Whoosh==2.7.4
wrapt==1.17.0
yarl==1.18.3
